[Feat] #17 - SemanticColor, ColorTheme Models 폴더로 파일 분리
