+++ Carmunication/Models/Color/ColorTheme.swift
@@ -1,0 +1,74 @@
+//
+//  ColorTheme.swift
+//  Carmunication
+//
+//  Created by 김동현 on 2023/10/01.
+//
+
+import UIKit
+/**
+ 기본 컬러 색상
+ */
+struct ColorTheme {
+    // MARK: Black & White
+    let black = UIColor(named: "Black")
+    let white = UIColor(named: "White")
+    let trans20 = UIColor(named: "Trans20")
+    let trans40 = UIColor(named: "Trans40")
+    // MARK: Gray Color
+    let gray1 = UIColor(named: "Gray1")
+    let gray2 = UIColor(named: "Gray2")
+    let gray3 = UIColor(named: "Gray3")
+    let gray4 = UIColor(named: "Gray4")
+    let gray5 = UIColor(named: "Gray5")
+    let gray6 = UIColor(named: "Gray6")
+    let gray7 = UIColor(named: "Gray7")
+    let gray8 = UIColor(named: "Gray8")
+    let gray9 = UIColor(named: "Gray9")
+    let gray10 = UIColor(named: "Gray10")
+    let gray11 = UIColor(named: "Gray11")
+    let gray12 = UIColor(named: "Gray12")
+    // MARK: Acua Color
+    let acua1 = UIColor(named: "Acua1")
+    let acua2 = UIColor(named: "Acua2")
+    let acua3 = UIColor(named: "Acua3")
+    let acua4 = UIColor(named: "Acua4")
+    let acua5 = UIColor(named: "Acua5")
+    let acua6 = UIColor(named: "Acua6")
+    let acua7 = UIColor(named: "Acua7")
+    let acua8 = UIColor(named: "Acua8")
+    let acua9 = UIColor(named: "Acua9")
+    let acuaTrans20 = UIColor(named: "AcuaTrans20")
+    // MARK: Blue Color
+    let blue1 = UIColor(named: "Blue1")
+    let blue2 = UIColor(named: "Blue2")
+    let blue3 = UIColor(named: "Blue3")
+    let blue4 = UIColor(named: "Blue4")
+    let blue5 = UIColor(named: "Blue5")
+    let blue6 = UIColor(named: "Blue6")
+    let blue7 = UIColor(named: "Blue7")
+    let blue8 = UIColor(named: "Blue8")
+    let blue9 = UIColor(named: "Blue9")
+    let blueTrans20 = UIColor(named: "BlueTrans20")
+    let blueTrans40 = UIColor(named: "BlueTrans40")
+    // MARK: Red Color
+    let red1 = UIColor(named: "Red1")
+    let red2 = UIColor(named: "Red2")
+    let red3 = UIColor(named: "Red3")
+    let red4 = UIColor(named: "Red4")
+    let red5 = UIColor(named: "Red5")
+    let red6 = UIColor(named: "Red6")
+    let red7 = UIColor(named: "Red7")
+    let red8 = UIColor(named: "Red8")
+    let red9 = UIColor(named: "Red9")
+    // MARK: Dark Blue
+    let darkblue1 = UIColor(named: "DarkBlue1")
+    let darkblue2 = UIColor(named: "DarkBlue2")
+    let darkblue3 = UIColor(named: "DarkBlue3")
+    let darkblue4 = UIColor(named: "DarkBlue4")
+    let darkblue5 = UIColor(named: "DarkBlue5")
+    let darkblue6 = UIColor(named: "DarkBlue6")
+    let darkblue7 = UIColor(named: "DarkBlue7")
+    let darkblue8 = UIColor(named: "DarkBlue8")
+    let darkblue9 = UIColor(named: "DarkBlue9")
+}
