+++ Carmunication/Models/Color/SemanticColor.swift
@@ -1,0 +1,25 @@
+//
+//  SemanticColor.swift
+//  Carmunication
+//
+//  Created by 김동현 on 2023/10/01.
+//
+
+import UIKit
+/**
+ 시멘틱 컬러 색상
+ */
+struct SemanticColor {
+    let backgroundDefault = UIColor.theme.white
+    let textPrimary = UIColor.theme.darkblue8
+    let textBody = UIColor.theme.darkblue8
+    let textSecondary = UIColor.theme.white
+    let backgroundSecond = UIColor.theme.blue1
+    let negative = UIColor.theme.red7
+    let error = UIColor.theme.red7
+    let accPrimary = UIColor.theme.blue6
+    let accSecondary = UIColor.theme.acua6
+    let backgroundDisableBT = UIColor.theme.blue1
+    let textDisableBT = UIColor.theme.blue3
+    let textNewCrew = UIColor.theme.red5
+}
